requirements.txt
Django==4.0.2
psycopg2==2.9.3
coverage==6.3.1
djangorestframework==3.13.1
django-filter==21.1
dj-rest-auth==2.2.3